// src/app/model/UserVO.ts
export class UserVO {
    userId?: string;
    userName?: string;
    userEmail?: string;
    pd?: string;
    success?: string;
    errMessage?: string;
    isAdmin?: string;
    managerName?: string;
    managerEmail?: string;
    proxyUserId?: string;
    hoUser?: string;
    isPurhaseUser?: string;
    isAPUser?: string;
    companyCode?: string;
    managerId?: string;
    costCenter?: string;
}