// src/app/model/CostCenterDetail.ts
export class CostCenterDetail {
    uname?: string;
    name?: string;
    email?: string;
    level?: string;
    frggr?: string;
    frgsx?: string;
    relCode?: string;
    mandt?: string;
    spras?: string;
    kokrs?: string;
    kostl?: string;
    datbi?: string;
    ktext?: string;
    ltext?: string;
    mctxt?: string;
}